feat: move you2be list to separate module
