+++ src/app/youtube/components/result-item/footer.directive.ts
