+++ src/app/youtube/components/result-item/result-item.component.scss
