+++ src/app/youtube/pages/results-list/results-list.component.scss
